facial_attendance/requirements.txt
flask
numpy
pillow
torch
facenet-pytorch
